server/article_service/requirements.txt
# HTTP Client
httpx>=0.24.0,<1.0.0

# RSS Parsing
feedparser>=6.0.10

# Web Scraping
trafilatura>=1.8.0
lxml>=4.9.0
lxml_html_clean>=0.1.0
beautifulsoup4>=4.11.0

# Content extraction
readability-lxml>=0.8.1

# Database
SQLAlchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.0

# Cache / Deduplication
redis[hiredis]>=5.0.0

# Scheduling
APScheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0

# Web Server
fastapi>=0.100.0
uvicorn>=0.22.0

# Messaging
kafka-python-ng>=2.2,<3.0
